transcription-demo-infra/pyproject.toml
[project]
name = "transcription-demo-infra"
version = "0.1.0"
description = "DevOps / IaC for transcription-demo (CDK: S3, Lambda, Bedrock)."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[tool.uv]
dev-dependencies = []
